约束无法删除默认助手

+++ blog/src/main/java/liuyuyang/net/web/controller/AssistantController.java
@@ -43,6 +43,10 @@ public class AssistantController {
     @ApiOperation("删除助手")
     @ApiOperationSupport(author = "刘宇阳 | liuyuyang1024@yeah.net", order = 2)
     public Result<String> del(@PathVariable Integer id) {
+        Assistant data = assistantService.getById(id);
+        if (data == null) return Result.error("该助手不存在");
+        if (data.getIsDefault() == 1) return Result.error("无法删除默认助手，请更换后重试");
+
         assistantService.removeById(id);
         return Result.success();
     }
